adjusts endpoints 2 e 6
